Fix v0.10.0 runtime environment
Previous build had a mistake due to how `uv` handles python binary
symlinks and its incompatibility with bugfix version bumps. This
resulted in the virtual environment, where vllm is installed, not being
able to find the python executable to actually be able to run vllm.

This updates the final build step to pull the necessary uv environment
files over from the build stage into the final runtime stage so
everything expected is in place.

Signed-off-by: Sean McGinnis <sean.mcginnis@gmail.com>